Arith: split std into std and arith
