+++ src/xdsl/dialects/arith.py
@@ -1,0 +1,215 @@
+from __future__ import annotations
+from xdsl.ir import *
+from xdsl.irdl import *
+from xdsl.util import *
+from xdsl.dialects.builtin import IntegerType, Float32Type, IntegerAttr, FlatSymbolRefAttr
+
+
+@dataclass
+class Arith:
+    ctx: MLContext
+
+    def __post_init__(self):
+        self.ctx.register_op(Constant)
+
+        self.ctx.register_op(Addi)
+        self.ctx.register_op(Muli)
+        self.ctx.register_op(Subi)
+        self.ctx.register_op(FloorDiviSI)
+        self.ctx.register_op(RemSI)
+
+        self.ctx.register_op(Addf)
+        self.ctx.register_op(Mulf)
+
+        self.ctx.register_op(Cmpi)
+
+
+@irdl_op_definition
+class Constant(Operation):
+    name: str = "arith.constant"
+    output = ResultDef(AnyAttr())
+    value = AttributeDef(AnyAttr())
+
+    # TODO verify that the output and value type are equal
+
+    @staticmethod
+    def from_attr(attr: Attribute, typ: Attribute) -> Constant:
+        return Constant.create(result_types=[typ], attributes={"value": attr})
+
+    @staticmethod
+    def from_int_constant(val: Union[int, Attribute],
+                          typ: Union[int, Attribute]) -> Constant:
+        if isinstance(typ, int):
+            typ = IntegerType.from_width(typ)
+        return Constant.create(
+            result_types=[typ],
+            attributes={"value": IntegerAttr.from_params(val, typ)})
+
+
+@irdl_op_definition
+class Addi(Operation):
+    name: str = "arith.addi"
+    input1 = OperandDef(IntegerType)
+    input2 = OperandDef(IntegerType)
+    output = ResultDef(IntegerType)
+
+    # TODO replace with trait
+    def verify_(self) -> None:
+        if self.input1.typ != self.input2.typ or self.input2.typ != self.output.typ:
+            raise Exception("expect all input and output types to be equal")
+
+    @staticmethod
+    def get(operand1: Union[Operation, SSAValue],
+            operand2: Union[Operation, SSAValue]) -> Addi:
+        return Addi.build(operands=[operand1, operand2],
+                          result_types=[IntegerType.from_width(32)])
+
+
+@irdl_op_definition
+class Muli(Operation):
+    name: str = "arith.muli"
+    input1 = OperandDef(IntegerType)
+    input2 = OperandDef(IntegerType)
+    output = ResultDef(IntegerType)
+
+    # TODO replace with trait
+    def verify_(self) -> None:
+        if self.input1.typ != self.input2.typ or self.input2.typ != self.output.typ:
+            raise Exception("expect all input and output types to be equal")
+
+    @staticmethod
+    def get(operand1: Union[Operation, SSAValue],
+            operand2: Union[Operation, SSAValue]) -> Muli:
+        return Muli.build(operands=[operand1, operand2],
+                          result_types=[IntegerType.from_width(32)])
+
+
+@irdl_op_definition
+class Subi(Operation):
+    name: str = "arith.subi"
+    input1 = OperandDef(IntegerType)
+    input2 = OperandDef(IntegerType)
+    output = ResultDef(IntegerType)
+
+    # TODO replace with trait
+    def verify_(self) -> None:
+        if self.input1.typ != self.input2.typ or self.input2.typ != self.output.typ:
+            raise Exception("expect all input and output types to be equal")
+
+    @staticmethod
+    def get(operand1: Union[Operation, SSAValue],
+            operand2: Union[Operation, SSAValue]) -> Subi:
+        return Subi.build(operands=[operand1, operand2],
+                          result_types=[IntegerType.from_width(32)])
+
+
+@irdl_op_definition
+class FloorDiviSI(Operation):
+    name: str = "arith.floordivsi"
+    input1 = OperandDef(IntegerType)
+    input2 = OperandDef(IntegerType)
+    output = ResultDef(IntegerType)
+
+    # TODO replace with trait
+    def verify_(self) -> None:
+        if self.input1.typ != self.input2.typ or self.input2.typ != self.output.typ:
+            raise Exception("expect all input and output types to be equal")
+
+    @staticmethod
+    def get(operand1: Union[Operation, SSAValue],
+            operand2: Union[Operation, SSAValue]) -> FloordiviSigned:
+        return FloordiviSigned.build(operands=[operand1, operand2],
+                                     result_types=[IntegerType.from_width(32)])
+
+
+@irdl_op_definition
+class RemSI(Operation):
+    name: str = "arith.remsi"
+    input1 = OperandDef(IntegerType)
+    input2 = OperandDef(IntegerType)
+    output = ResultDef(IntegerType)
+
+    # TODO replace with trait
+    def verify_(self) -> None:
+        if self.input1.typ != self.input2.typ or self.input2.typ != self.output.typ:
+            raise Exception("expect all input and output types to be equal")
+
+    @staticmethod
+    def get(operand1: Union[Operation, SSAValue],
+            operand2: Union[Operation, SSAValue]) -> RemiSigned:
+        return RemiSigned.build(operands=[operand1, operand2],
+                                result_types=[IntegerType.from_width(32)])
+
+
+@irdl_op_definition
+class And(Operation):
+    name: str = "arith.and"
+    input1 = OperandDef(IntegerType)
+    input2 = OperandDef(IntegerType)
+    output = ResultDef(IntegerType)
+
+    # TODO replace with trait
+    def verify_(self) -> None:
+        if self.input1.typ != self.input2.typ or self.input2.typ != self.output.typ:
+            raise Exception("expect all input and output types to be equal")
+
+    @staticmethod
+    def get(operand1: Union[Operation, SSAValue],
+            operand2: Union[Operation, SSAValue]) -> And:
+        return And.build(operands=[operand1, operand2],
+                         result_types=[IntegerType.from_width(1)])
+
+
+@irdl_op_definition
+class Cmpi(Operation):
+    name: str = "arith.cmpi"
+    predicate = AttributeDef(IntegerAttr)
+    input1 = OperandDef(IntegerType)
+    input2 = OperandDef(IntegerType)
+    output = ResultDef(IntegerType.from_width(1))
+
+    @staticmethod
+    def get(operand1: Union[Operation, SSAValue],
+            operand2: Union[Operation, SSAValue], arg: int) -> Cmpi:
+        return Cmpi.build(
+            operands=[operand1, operand2],
+            result_types=[IntegerType.from_width(1)],
+            attributes={"predicate": IntegerAttr.from_int_and_width(arg, 64)})
+
+
+@irdl_op_definition
+class Addf(Operation):
+    name: str = "arith.addf"
+    input1 = OperandDef(Float32Type)
+    input2 = OperandDef(Float32Type)
+    output = ResultDef(Float32Type)
+
+    # TODO replace with trait
+    def verify_(self) -> None:
+        if self.input1.typ != self.input2.typ or self.input2.typ != self.output.typ:
+            raise Exception("expect all input and output types to be equal")
+
+    @staticmethod
+    def get(operand1: Union[Operation, SSAValue],
+            operand2: Union[Operation, SSAValue]) -> Addf:
+        return Addf.build(operands=[operand1, operand2],
+                          result_types=[Float32Type()])
+
+
+@irdl_op_definition
+class Mulf(Operation):
+    name: str = "arith.mulf"
+    input1 = OperandDef(Float32Type)
+    input2 = OperandDef(Float32Type)
+    output = ResultDef(Float32Type)
+
+    # TODO replace with trait
+    def verify_(self) -> None:
+        if self.input1.typ != self.input2.typ or self.input2.typ != self.output.typ:
+            raise Exception("expect all input and output types to be equal")
+
+    @staticmethod
+    def get(operand1: Union[Operation, SSAValue],
+            operand2: Union[Operation, SSAValue]) -> Mulf:
+        return Mulf.build(operands=[operand1, operand2],
+                          result_types=[Float32Type()])
