--- tests/filecheck/std_ops.test
@@ -1,23 +1,0 @@
-// RUN: xdsl-opt -f %s | xdsl-opt | filecheck %s
-
-module() {
-  builtin.func() ["sym_name" = "test", "type" = !fun<[], []>, "sym_visibility" = "private"]{
-    %0 : !i32 = std.constant() ["value" = 42 : !i32]
-    %1 : !i32 = std.constant() ["value" = 42 : !i32]
-    %2 : !i32 = std.addi(%0 : !i32, %1 : !i32) 
-    %3 : !i32 = std.subi(%0 : !i32, %1 : !i32) 
-    %4 : !i32 = std.muli(%0 : !i32, %1 : !i32) 
-    %5 : !i32 = std.floordivi_signed(%0 : !i32, %1 : !i32) 
-    %6 : !i32 = std.remi_signed(%0 : !i32, %1 : !i32) 
-  }
-}
-
-// CHECK-LABEL: builtin.func() ["sym_name" = "test"
-// CHECK: %{{.*}} : !i32 = std.constant() ["value" = 42 : !i32] 
-// CHECK: %{{.*}} : !i32 = std.constant() ["value" = 42 : !i32]
-// CHECK: %{{.*}} : !i32 = std.addi(%{{.*}} : !i32, %{{.*}} : !i32) 
-// CHECK: %{{.*}} : !i32 = std.subi(%{{.*}} : !i32, %{{.*}} : !i32) 
-// CHECK: %{{.*}} : !i32 = std.muli(%{{.*}} : !i32, %{{.*}} : !i32) 
-// CHECK: %{{.*}} : !i32 = std.floordivi_signed(%{{.*}} : !i32, %{{.*}} : !i32) 
-// CHECK: %{{.*}} : !i32 = std.remi_signed(%{{.*}} : !i32, %{{.*}} : !i32) 
-
